🔨 refactor: 재사용되는 타입을 types/shared로 분리
