+++ app/types/shared/error.d.ts
@@ -1,0 +1,3 @@
+export interface ErrorResponse {
+  message: string;
+}
