+++ app/types/shared/group.d.ts
@@ -1,0 +1,8 @@
+export interface GroupType {
+  teamId: string;
+  updatedAt: string;
+  createdAt: string;
+  image: string;
+  name: string;
+  id: number;
+}
