+++ app/types/user.d.ts
@@ -1,18 +1,4 @@
-import { UserType } from './auth';
-import { GroupType } from './group';
-import { TaskHistoryType } from './task';
-
-type Role = 'ADMIN' | 'MEMBER';
-
-export interface MemberShipType {
-  group: GroupType;
-  role: Role;
-  userImage: string;
-  userEmail: string;
-  userName: string;
-  groupId: number;
-  userId: number;
-}
+import { GroupType, MemberShipType, TaskHistoryType, UserType } from './shared';
 
 export interface UserDataType extends UserType {
   memberships: MemberShipType[];
